requirements/base.txt
appdirs==1.4.3
bleach==2.0.0
Django==1.10.6
django-braces==1.11.0
django-markdownify==0.1.0
django-tagulous==0.12.0
html5lib==0.999999999
Markdown==2.6.8
olefile==0.44
packaging==16.8
Pillow==4.0.0
pyparsing==2.2.0
six==1.10.0
webencodings==0.5
python-dotenv
black